% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_filter.R
\name{adaptive_filter}
\alias{adaptive_filter}
\title{Perform adaptive filter}
\usage{
adaptive_filter(signal, base_order, adaptivity_factor)
}
\arguments{
\item{signal}{Incoming signal}

\item{base_order}{Base order of Laguerre's polynomial series}

\item{adaptivity_factor}{A parameter that controls the degree of variation in the order
of the Laguerre filter as a function of the local variance of the signal.}
}
\value{
filtered_signal
}
\description{
This function applies  adaptive filter on a signal using Laguerre polynomials
}
\examples{
signal <- c(1, 2, 3, 4, 5, 4, 3, 2, 1)
base_order <- 3
adaptivity_factor <- 5
filtered_signal <- adaptive_filter(signal, base_order, adaptivity_factor)
}
