% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Name drawer}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{name to be ploted}

\item{the_sex}{sex of the same}
}
\value{
plot
}
\description{
plots the name popularity
}
