% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Name comparator}
\usage{
draw_names(names)
}
\arguments{
\item{names}{vector of the names to be}
}
\value{
plot of the 2 names
}
\description{
Compares the popularity of 2 names
}
