% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laguerre_poly.R
\name{Laguerre_poly}
\alias{Laguerre_poly}
\title{Calculate Laguerre polynomial of order n}
\usage{
Laguerre_poly(order, x)
}
\arguments{
\item{order}{Order of Laguerre polynomial}

\item{x}{Argument of Laguerre polynomial}
}
\value{
Laguerre polynomial of order n
}
\description{
This function calculate Laguerre polynomial of order n
}
\examples{
Laguerre_poly(5,1)
}
