% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myc_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{csv Reader}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{containing csv files to be read}
}
\value{
list of files from the folder
}
\description{
csv Reader
}
