% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_multi.R
\name{read_excel_multi}
\alias{read_excel_multi}
\title{Read excel file}
\usage{
read_excel_multi(file)
}
\arguments{
\item{file}{excel file to be used}
}
\value{
read file
}
\description{
Read excel file
}
