% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{csv saver}
\usage{
save_as_csv(dataset, filename, row.names = FALSE)
}
\arguments{
\item{dataset}{data set to be saved}

\item{filename}{name to give it}

\item{row.names}{extra}
}
\value{
saved file
}
\description{
Saves a data set
}
